1.[MOVE]:move the network status to def.hpp, so that it could be used bt both http and tcp
